adding ajax request for update and delete of post

+++ controllers/post_controller.js
@@ -3,10 +3,20 @@ const comment = require('../models/comment')
 
 module.exports.create = async function (req, res) {
 try{
-    let posts = await Post.create({
+    let post = await Post.create({
         content: req.body.content,
         user: req.user._id
     });
+
+    if(req.xhr){
+        return res.status(200).json({
+            data:{
+                post:post
+            },
+            message:"post created"
+            
+        })
+    }
 
         return res.redirect('back');
 }
@@ -26,7 +36,15 @@ module.exports.destroy = async function (req, res) {
             console.log(post.user)
             post.remove();
 
-            let comments = await comment.deleteMany({ post: req.params.id })
+            if(req.xhr){
+                return res.status(200).json({
+                    data : {
+                    post_id: req.params.id
+                  },
+                  message : "Post deleted"
+                });
+            }
+            await comment.deleteMany({ post: req.params.id })
                 return res.redirect('back');
         
         }
